Fixed #2 and #5 - Sculpt Settings Persistence
Issue #5 was reported by me (Aldrin Mathew) and is a superset of #2.
This commit fixes that by storing all the non persistent mesh data and then reassigning it to the newly created mesh.
Now Voxel Size, Voxel Adaptivity, Symmetry Mirror options in the Sculpt Mode settings will be the same as it was before the keyframe was added.