Added credits
Since I am fixing bugs in the project,
I have credited myself as "Aldrin Mathew (Improvements)",
in addition to the previous contributors

If there is any issue about this, I will gladly remove my name from the project.
It's just that this is my first Python project, and it is not my favourite language.
So I have to put in extra effort to make improvements to the addon.